requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
requests==2.28.2
python-dotenv==1.0.0
langchain==0.0.235
pydantic==1.10.7
pytest==7.3.1
httpx==0.24.0
openai==0.27.0